requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pytest>=8.0.0
tqdm>=4.66.0
pdfplumber>=0.10.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
openai>=1.50.0
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
